+++ src/evaluate.ts
@@ -240,15 +240,26 @@ function evaluateEnclosingForm(document = {}, options = {}) {
     })).catch(printWarningForError);
 }
 
-
-function evaluateToCursor(document = {}, options = {}) {
+function evaluateUsingTextAndSelectionGetter(getter: (editor: vscode.TextEditor) => getText.SelectionAndText, formatter: (s: string) => string, document = {}, options = {}) {
     evaluateSelection(document, Object.assign({}, options, {
         pprintOptions: getConfig().prettyPrintingOptions,
         selectionFn: (editor: vscode.TextEditor) => {
-            let [selection, code] = getText.toStartOfList(editor);
-            return [selection, `(${code})`]
+            let [selection, code] = getter(editor);
+            return [selection, formatter(code)]
         }
     })).catch(printWarningForError);
+}
+
+function evaluateToCursor(document = {}, options = {}) {
+    evaluateUsingTextAndSelectionGetter(getText.toStartOfList, code => `(${code})`, document, options);
+}
+
+function evaluateTopLevelFormToCursor(document = {}, options = {}) {
+    evaluateUsingTextAndSelectionGetter(getText.currentTopLevelFormToCursor, code => `${code}`, document, options);
+}
+
+function evaluateStartOfFileToCursor(document = {}, options = {}) {
+    evaluateUsingTextAndSelectionGetter(getText.startOFileToCursor, code => `${code}`, document, options);
 }
 
 async function loadFile(document, pprintOptions: PrettyPrintingOptions) {
@@ -407,6 +418,8 @@ export default {
     evaluateSelectionAsComment,
     evaluateTopLevelFormAsComment,
     evaluateToCursor,
+    evaluateTopLevelFormToCursor,
+    evaluateStartOfFileToCursor,
     evaluateCode,
     evaluateUser,
     copyLastResultCommand,